vayu/api/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
azure-storage-blob==12.22.0
certifi==2021.10.8
chardet==4.0.0
distlib==0.3.4
Django==3.1.5
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.13.0
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.7.0
django-import-export==2.7.0
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.5
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
Pillow==9.0.0
psycopg2-binary==2.9.5
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.25.1
six==1.16.0 
sqlparse==0.2.4
urllib3==1.26.7
virtualenv==20.4.2
